👌 Clean up call structure
