+++ upgrades/0.0.94.persist-loki/pkg/lib/s3/types.go
@@ -1,3 +1,7 @@
 package s3
 
-const defaultStackTimeoutMinutes = 5
+const (
+	defaultStackTimeoutMinutes = 5
+	defaultLogicalBucketName   = "S3Bucket"
+	defaultBucketARNOutputName = "BucketARN"
+)
